Create zig_zag_sequence.py
The function wasn't defining the mid of the array properly since it's wasn't counting the first element is 0 and not 1 so mid was defined as before less 1. Then the variable ed has to be lesser by one (started in that point) since whe want to make the previous mid element in the last sin it's going to be the lesser number for the second sequeuence k. Then fixed ed = ed+1 error since ed can be higher than "n-1" since it's the last number in the array.